deprecated-backends/ml-backend/Flask/requirements.txt
Flask
opencv-python
tensorflow
Pillow
numpy
Werkzeug